% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bird_dictionary}
\alias{bird_dictionary}
\title{Bird dictionary}
\format{
A data frame.
}
\usage{
bird_dictionary
}
\description{
Dictionary linking eBird scientific names with common names and status.
}
\keyword{datasets}
