% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_map.R
\name{bird_map}
\alias{bird_map}
\title{Create a leaflet map from eBird observations}
\usage{
bird_map(ebird_file)
}
\arguments{
\item{ebird_file}{Path to a MyEBirdData CSV file.}
}
\value{
A leaflet map.
}
\description{
Create a leaflet map from eBird observations
}
\examples{
\dontrun{
bird_map("data/MyEBirdData.csv")
}
}
