% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MyEBirdData}
\alias{MyEBirdData}
\title{Example eBird dataset}
\format{
A data frame.
}
\usage{
MyEBirdData
}
\description{
Example MyEBirdData file used to illustrate the package.
}
\keyword{datasets}
